requirements.txt
Flask-RESTful
Flask-SQLAlchemy
git+ssh://git@github.com/gastongr/FuelSDK-Python.git@master#egg=fuelsdk
